refactored tests to move excerpt test code in own class
